+++ tests/testthat/fakepackages/notallexportedchecked/R/pending.R
@@ -1,0 +1,12 @@
+#' Pending!
+#'
+#' This function has no formals, so it won't be counted against checkr!
+#' @export
+pending <- function() { "Pending!" }
+
+
+#' Pending identity.
+#'
+#' This function needs to be checked or else the test will fail. But it's not checked!
+#' @export
+pending_identity <- identity